+++ modelcif/__init__.py
@@ -4,7 +4,7 @@ from ihm import AsymUnitRange, _remove_identical  # noqa: F401
 import modelcif.data
 import sys
 
-__version__ = '0.3'
+__version__ = '0.4'
 
 
 class System(object):